Refactor conch functionality into `single_leader`
Factor out reliable, single-leader functionality

Fix `single_leader` include error
